refactor(landcover): the classifier, its reference and the legend both draw in
Five modules move into terra/landcover/: mapbiomas (was lulc), palette (was
class_palette), and domain_shift, prithvi and temporal_transformer unchanged.

lulc is an acronym for land use and land cover, which is what the package is
now called, so inside it the name says the package twice and says nothing else.
What the module actually provides is the MapBiomas reference: the raster the
classification is compared against and the shares it reports. class_palette is
palette once the package says which classes.

The palette module records why it exists, and it belongs here for that reason:
the classification path and the reference path each carried their own colour
literal, the two disagreed on every class they shared, and a prediction and the
MapBiomas map it is compared against came out painted differently for the same
class. Both readers are in this slice; the module goes with them rather than
below them.

The contract now names five products and the independence between them is
checked: land cover, water, flood, canopy and energy do not import each other.

sidecar/ holds two Python files that are not the package. infer.py is the entry
point the application runs; doctor.py is a second one, run in the CANDIDATE
interpreter rather than the shipped one, to decide whether that environment can
run the sidecar at all. internal/pyenv builds its path and a test fails if the
name changes. Both stay where the Go side expects to find them.

Verification: pytest 517 passed, 8 skipped. ruff check clean, three contracts
kept, gofmt, go vet and go test -race clean across every package. A ping
through the entry point answers.
