+++ sidecar/pyproject.toml
@@ -72,7 +72,7 @@ layers = [
     "terra.cli",
     "terra.registry",
     "terra.actions",
-    "terra.energy | terra.water | terra.flood | terra.canopy",
+    "terra.energy | terra.water | terra.flood | terra.canopy | terra.landcover",
     "terra.imagery | terra.terrain | terra.sun",
     "terra.stac",
     "terra.protocol",
@@ -81,7 +81,13 @@ layers = [
 [[tool.importlinter.contracts]]
 name = "Products do not reach into each other"
 type = "independence"
-modules = ["terra.energy", "terra.water", "terra.flood", "terra.canopy"]
+modules = [
+    "terra.energy",
+    "terra.water",
+    "terra.flood",
+    "terra.canopy",
+    "terra.landcover",
+]
 
 [[tool.importlinter.contracts]]
 name = "The sun service knows nothing about the products that read it"
