+++ sidecar/terra/landcover/__init__.py
@@ -1,0 +1,14 @@
+"""
+What is growing on the ground, and how wrong the answer is.
+
+Three model paths over one AOI protocol, all emitting the same MapBiomas
+classes; the reference those classes are compared against; the palette both
+sides are drawn in; and the measurement of how far a run sits from the ground
+the models were fitted on.
+
+THE PALETTE IS HERE FOR A REASON IT RECORDS ITSELF. The classification path and
+the reference path each carried their own colour literal, the two disagreed on
+every class they shared, and a prediction and the MapBiomas map it is compared
+against came out painted differently for the same class. One module, one
+legend, both sides.
+"""
