+++ sidecar/terra/landcover/palette.py
