+++ sidecar/terra/landcover/prithvi.py
