+++ sidecar/terra/landcover/temporal_transformer.py
